# logsumexp for multiple dimensions