@umanwizard 2019-01-29
I think the higher time for `dim=1` on CPU was a fluke. I ran it again and got 237 ms.